MacOSX tries to use java.awt.Desktop if it's running java 1.6 or 1.7 if it ever comes out, fixes issue with jdic.
The rest will try to use java.awt.Desktop or the old GUIMediator way of launching files.

Graphics library ads a 512x512 frostwire sphere icon, and another huge one. The 512x512 came to existence since the android market asked for it, we'll keep it in the repo if anyone needs it.