Some work towards an easier to understand user experience "Finished Downloads" directory holder.
The idea is to hide all folders from the UI, since it frustrates the user that wants to use it as a file explorer and we're trying to get away from that metaphor.

Added a cache on SavedFilesDirectoryHolder, but I'm thinking of scraping that and understanding how the caches for the other directory holders work and use that mechanism instead.