Only automatically resume those transfers that were active when the connection was dropped.
When we pause the VuzeManager, now we pass a 'disconnected' flag. If we are stopping due to disconnection, then those torrents that are not already 'STOPPED' will be marked as 'QUEUED'.
The Vuze documentation says the QUEUED state is precisely for this purpose. QUEUED is the same as STOPPED but with the only difference that it can be restarted automatically.