Fixes issues when opening FrostWire out of magnet or .torrent file on disk.
FrostWire would freeze when opening from a magnet url due to a race condition,
and it would not add a .torrent to the transfer manager if invoked from a file
manager by opening such .torrent file.

Now FrostWire waits for the Azureus core to be started and for any previous transfers
to be started before attempting to fetch a magnet or open a .torrent file for download.