use original_file_format from soundcloud to put the file extension to a soundcloud download.
this was an attempt to fix issue #211. Renaming the file to the right file extension does not work.
It seems as if the .wav file sent by soundcloud comes with an extra header.
I also tried chopping off the beginning extra bytes of the downloaded file, but the file was still
not playable, not sure then what it is that we download when the downloadable file isn't an mp3.

won't fix.